app/requirements.txt
boto3
dataclasses
Flask
paho-mqtt